fix: do not try to map anime with empty string
This commit fixes an issue where the mal api gives use a synonym as an
empty string (it should normally return null). If trying to map to the
provider it can happen that the empty string triggers a download of the
whole anime catalouge of the provider, we do not want this!